+++ reikna/core/signature.py
@@ -100,6 +100,21 @@ class Type:
 
         return True
 
+    def broadcastable_to(self, other):
+        """
+        Returns ``True`` if the shape of this ``Type`` is broadcastable to ``other``,
+        that is its dimensions either coincide with the innermost dimensions of ``other.shape``,
+        or are equal to 1.
+        """
+        if len(self.shape) > len(other.shape):
+            return False
+
+        for i in range(1, len(self.shape)+1):
+            if not (self.shape[-i] == 1 or self.shape[-i] == other.shape[-i]):
+                return False
+
+        return True
+
     def with_dtype(self, dtype):
         """
         Creates a :py:class:`Type` object with its ``dtype`` attribute replaced by the given dtype.